+++ examples/lastrev-next-starter/packages/components/src/components/Media/Media.mock.ts
@@ -3,9 +3,9 @@ import { MediaProps } from './Media';
 
 export const mediaMock = (): MediaProps => ({
   file: {
-    url: './flower-large.jpg',
-    width: '920',
-    height: '613'
+    url: '/flower-large.jpg',
+    width: 920,
+    height: 613
   },
   title: lorem.sentence(),
   description: lorem.sentence()
@@ -14,8 +14,8 @@ export const mediaMock = (): MediaProps => ({
 export const mediaVideoMock = () => ({
   file: {
     url: './LastRev.mp4',
-    width: '1280',
-    height: '720'
+    width: 1280,
+    height: 720
   },
   variant: 'video',
   title: lorem.sentence(),
@@ -32,29 +32,29 @@ export const assetMock = () => ({
 
 export const fileMock = () => ({
   url: `https://testImage-${Date.now().toString()}-${lorem.word()}-${lorem.word()}/cmp.png?h=180&r=180`,
-  width: '180',
-  height: '180'
+  width: 180,
+  height: 180
 });
 
 export const responsiveMediaMock = {
   __typename: 'Media',
   file: {
     // url: flowerLarge,
-    url: './flower-large.jpg',
-    width: '1728px',
-    height: '1152px'
+    url: '/flower-large.jpg',
+    width: 1728,
+    height: 1152
   },
   fileTablet: {
     // url: flowerMedium,
     url: './flower-medium.jpg',
-    width: '920',
-    height: '613'
+    width: 920,
+    height: 613
   },
   fileMobile: {
     // url: flowerSmall,
     url: './flower-small.jpg',
-    width: '540',
-    height: '540'
+    width: 540,
+    height: 540
   },
   title: lorem.sentence(),
   description: lorem.sentence()
@@ -65,8 +65,8 @@ export const SVGMediaMock = {
   file: {
     // url: flowerLarge,
     url: './logo.svg',
-    width: '1728px',
-    height: '1152px'
+    width: 1728,
+    height: 1152
   },
 
   title: lorem.sentence(),
@@ -78,8 +78,8 @@ export const ExternalSVGMediaMock = {
   file: {
     // url: flowerLarge,
     url: './logo.svg',
-    width: '1728px',
-    height: '1152px'
+    width: 1728,
+    height: 1152
   },
   disableInlineSVG: true,
   title: lorem.sentence(),